+++ src/StickyTabs/index.tsx
@@ -14,9 +14,10 @@ interface IProps {
   tabL: ITabItem[]; // tab列表
   offsetTop?: number; // 容器距顶fixed的top距离
   headEle?: React.ReactNode; // 顶部包裹固定的元素
-  tabBarExtraContent?: any;
-  // scrollContainer是为了兼容容器是scroll的，传入容器元素的处理目前可能存在问题（要求容器顶部区域为固定高度区域，不滚动！！！）
-  // 若不是必须先不考虑scrollContainer
+  tabBarExtraContent?: any; // tabs条右侧内容
+  // 副作用：smooth形式下点击tab滚动，短时间内可能会导致自行滚动的tab active无效
+  behavior?: 'instant' | 'smooth'; // 滚动形式1.instant 表示滚动会直接跳转到目标位置 2.smooth 平滑滚动并产生过渡效果
+  // scrollContainer设置滚动的容器，默认是window
   scrollContainer?: Element;
 }
 
@@ -24,6 +25,21 @@ export interface ITabItem {
   key: string;
   label: string | React.ReactNode;
   children?: React.ReactNode;
+}
+
+function shallowEqual(arr1: ITabItem[], arr2: ITabItem[]) {
+  // 检查属性名的数量是否相同
+  if (arr1.length !== arr2.length) {
+    return false;
+  }
+  let flag = true;
+  arr1.forEach((item, idx) => {
+    if (item.key !== arr2[idx].key) {
+      flag = false;
+    }
+  });
+
+  return flag;
 }
 
 /**
@@ -39,23 +55,34 @@ const StickyTabs = (props: IProps) => {
     scrollContainer,
     offsetTop = 0,
     tabBarExtraContent,
+    behavior = 'instant',
   } = props;
   const sc = (scrollContainer || window) as any; // 容器
+
   const headRef = useRef<any>(null); // head元素
   const idleCallbackRef = useRef<any>(null);
   const childRefs = useRef<any>({}); // tabs内容元素
   const stickyComp = useRef<any>(null); // 组件元素
+  const timer = useRef<any>(null);
   const [activeKey, setActiveKey] = useState<string>(); // active的tab
   const [containerResizeListener, containerSizes] = useResizeAware(); // 监听head高度变化hook，需要重新计算headH
   const [headH, setHeadH] = useState<number>(0); // 固定的head高度，相对于视口
   const [distance, setDistance] = useState<number>(0); // 当前组件距离固定容器的距离
+  const [isPActive, setIsPActive] = useState<boolean>(true);
+  const [tabHeads, setTabHeads] = useState<ITabItem[]>([]);
 
-  const tabHeads = useMemo(() => {
-    return tabL.map((item) => {
+  useEffect(() => {
+    const _tabHeads = tabL.map((item) => {
       const { label, key } = item;
       return { label, key };
     });
+    setTabHeads(_tabHeads);
+
+    if (!shallowEqual(_tabHeads, tabHeads)) {
+      setActiveKey((_tabHeads || [])[0]?.key);
+    }
   }, [tabL]);
+
   useEffect(() => {
     // 获取fixed顶部元素高度，顶部元素会吸顶，滚动时需要去掉
     if (sc) {
@@ -94,25 +121,48 @@ const StickyTabs = (props: IProps) => {
       setDistance(extraHeight);
     }
 
-    idleCallbackRef.current = requestIdleCallback(calculateDistanceToTop);
+    if ('requestIdleCallback' in window) {
+      idleCallbackRef.current = requestIdleCallback(calculateDistanceToTop);
+    } else {
+      idleCallbackRef.current = setTimeout(calculateDistanceToTop);
+    }
   };
 
   // 因为顶部距离不一定是固定的，所以需要监听获取
   useEffect(() => {
-    idleCallbackRef.current = requestIdleCallback(calculateDistanceToTop);
+    if ('requestIdleCallback' in window) {
+      idleCallbackRef.current = requestIdleCallback(calculateDistanceToTop);
+    } else {
+      idleCallbackRef.current = setTimeout(calculateDistanceToTop);
+    }
 
     return () => {
-      cancelIdleCallback(idleCallbackRef.current);
+      if ('requestIdleCallback' in window) {
+        cancelIdleCallback(idleCallbackRef.current);
+      } else {
+        clearTimeout(idleCallbackRef.current);
+      }
     };
   }, [distance, stickyComp, sc]);
 
+  const getScPaddingTop = () => {
+    if (!sc?.getBoundingClientRect) {
+      return 0;
+    }
+    return +window.getComputedStyle(sc).paddingTop.replace('px', '');
+  };
+
   const scrollCb = useCallback(() => {
+    if (!isPActive) {
+      return;
+    }
     let _activeKey;
+    // 吸顶后的顶部固定高度
     let _h = headH + offsetTop;
     if (sc?.getBoundingClientRect) {
       // 默认容器外是固定的
       const { top: containerOffsetTop = 0 } = sc?.getBoundingClientRect();
-      _h += containerOffsetTop;
+      _h = _h + containerOffsetTop + getScPaddingTop();
     }
     // 吸顶的时候，关注的是head高度还有固定容器距顶高度
     Object.keys(childRefs.current).forEach((key) => {
@@ -131,7 +181,7 @@ const StickyTabs = (props: IProps) => {
     if (_activeKey) {
       setActiveKey(_activeKey);
     }
-  }, [setActiveKey, childRefs.current, headH, distance]);
+  }, [setActiveKey, childRefs.current, headH, distance, sc, isPActive]);
 
   // 获取所有的tab元素
   useEffect(() => {
@@ -151,12 +201,32 @@ const StickyTabs = (props: IProps) => {
   }, [scrollCb]);
 
   const onChange = (key: string) => {
+    // 元素到stickyTabs的顶部距离 + stickyTabs距离包裹容器的距离 - stickyTabs的headH（sticky） - fixed距顶距离 - 滚动容器paddingTop（sticky附着在padding下） + 1
+    // + 1 是因为计算存在偏差，四舍五入导致的误差
+    const scrollTop =
+      childRefs.current[key].offsetTop +
+      distance -
+      headH -
+      offsetTop +
+      1 -
+      getScPaddingTop();
     sc.scrollTo({
-      // 元素到stickyTabs的顶部距离 + stickyTabs距离包裹容器的距离 - stickyTabs的headH（sticky） - fixed距顶距离 + 1
-      // + 1 是因为计算存在偏差，四舍五入导致的误差
-      top: childRefs.current[key].offsetTop + distance - headH - offsetTop + 1,
-      behavior: 'smooth',
+      top: scrollTop,
+      behavior,
     });
+    // 滚不到激活tab，需要自己行动，禁用一下position active
+    setActiveKey(key);
+    setIsPActive(false);
+    if (timer.current) {
+      clearTimeout(timer.current);
+    }
+    timer.current = setTimeout(
+      () => {
+        setIsPActive(true);
+        timer.current = null;
+      },
+      behavior === 'instant' ? 500 : 1500,
+    );
   };
 
   return (
@@ -178,6 +248,9 @@ const StickyTabs = (props: IProps) => {
           <Tabs
             tabBarExtraContent={tabBarExtraContent}
             activeKey={activeKey}
+            onFocus={(e) => {
+              e.target.blur();
+            }}
             onChange={onChange}
           >
             {(tabHeads || []).map((item) => {